+++ lib/widget/gridView.dart
@@ -4,11 +4,20 @@ import 'package:niku/macros/macros.dart';
 
 import '../objects/objects.dart';
 
+enum NikuGridViewType {
+  builder,
+  count,
+  custom,
+  extent,
+}
+
 // ignore: must_be_immutable
 class NikuGridView extends StatelessWidget
     with NikuBuildMacro, UseQueryMacro<NikuGridView>, ClipMacro, PaddingMacro {
-  List<Widget> children;
-  int crossAxisCount;
+  NikuGridViewType? type = NikuGridViewType.count;
+
+  List<Widget>? children;
+  int? itemCount;
 
   Axis? scrollDirection;
   bool? reverse;
@@ -30,10 +39,18 @@ class NikuGridView extends StatelessWidget
   String? restorationId;
   Clip? clipBehavior;
 
-  NikuGridView(
-    this.crossAxisCount,
-    this.children, {
+  IndexedWidgetBuilder? itemBuilder;
+  double? maxCrossExtent;
+  double? maxCrossAxisExtent;
+
+  SliverGridDelegate? gridDelegate;
+  SliverChildBuilderDelegate? childrenDelegate;
+
+  NikuGridView({
     Key? key,
+    this.type,
+    this.itemCount,
+    this.children,
     this.scrollDirection,
     this.reverse,
     this.controller,
@@ -53,14 +70,232 @@ class NikuGridView extends StatelessWidget
     this.keyboardDismissBehavior,
     this.restorationId,
     this.clipBehavior,
+    this.itemBuilder,
+    this.maxCrossAxisExtent,
+    this.maxCrossExtent,
+    this.gridDelegate,
+    this.childrenDelegate,
   }) : super(key: key);
 
-  set count(int value) => this.crossAxisCount = value;
+  factory NikuGridView.count({
+    List<Widget> children = const [],
+    int? crossAxisCount,
+    Key? key,
+    Axis? scrollDirection,
+    bool? reverse,
+    ScrollController? controller,
+    bool? primary,
+    ScrollPhysics? physics,
+    bool? shrinkWrap,
+    EdgeInsetsGeometry? padding,
+    double? mainAxisSpacing,
+    double? crossAxisSpacing,
+    double? childAspectRatio,
+    bool? addAutomaticKeepAlives,
+    bool? addRepaintBoundaries,
+    bool? addSemanticIndexes,
+    double? cacheExtent,
+    int? semanticChildCount,
+    DragStartBehavior? dragStartBehavior,
+    ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior,
+    String? restorationId,
+    Clip? clipBehavior,
+  }) =>
+      NikuGridView(
+        type: NikuGridViewType.count,
+        key: key,
+        itemCount: crossAxisCount ?? 1,
+        children: children,
+        scrollDirection: scrollDirection,
+        reverse: reverse,
+        controller: controller,
+        primary: primary,
+        physics: physics,
+        shrinkWrap: shrinkWrap,
+        padding: padding?.asNiku,
+        mainAxisSpacing: mainAxisSpacing,
+        crossAxisSpacing: crossAxisSpacing,
+        childAspectRatio: childAspectRatio,
+        addAutomaticKeepAlives: addAutomaticKeepAlives,
+        addRepaintBoundaries: addRepaintBoundaries,
+        addSemanticIndexes: addSemanticIndexes,
+        cacheExtent: cacheExtent,
+        semanticChildCount: semanticChildCount,
+        dragStartBehavior: dragStartBehavior,
+        keyboardDismissBehavior: keyboardDismissBehavior,
+        restorationId: restorationId,
+        clipBehavior: clipBehavior,
+      );
+
+  factory NikuGridView.builder({
+    Key? key,
+    required IndexedWidgetBuilder itemBuilder,
+    required SliverGridDelegate? gridDelegate,
+    int? itemCount,
+    Axis? scrollDirection,
+    bool? reverse,
+    ScrollController? controller,
+    bool? primary,
+    ScrollPhysics? physics,
+    bool? shrinkWrap,
+    NikuEdgeInsets? padding,
+    double? mainAxisSpacing,
+    double? crossAxisSpacing,
+    double? childAspectRatio,
+    bool? addAutomaticKeepAlives,
+    bool? addRepaintBoundaries,
+    bool? addSemanticIndexes,
+    double? cacheExtent,
+    int? semanticChildCount,
+    DragStartBehavior? dragStartBehavior,
+    ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior,
+    String? restorationId,
+    Clip? clipBehavior,
+    double? maxCrossExtent,
+  }) =>
+      NikuGridView(
+        type: NikuGridViewType.builder,
+        key: key,
+        itemCount: itemCount,
+        scrollDirection: scrollDirection,
+        reverse: reverse,
+        controller: controller,
+        primary: primary,
+        physics: physics,
+        shrinkWrap: shrinkWrap,
+        padding: padding,
+        mainAxisSpacing: mainAxisSpacing,
+        crossAxisSpacing: crossAxisSpacing,
+        childAspectRatio: childAspectRatio,
+        addAutomaticKeepAlives: addAutomaticKeepAlives,
+        addRepaintBoundaries: addRepaintBoundaries,
+        addSemanticIndexes: addSemanticIndexes,
+        cacheExtent: cacheExtent,
+        semanticChildCount: semanticChildCount,
+        dragStartBehavior: dragStartBehavior,
+        keyboardDismissBehavior: keyboardDismissBehavior,
+        restorationId: restorationId,
+        clipBehavior: clipBehavior,
+        maxCrossExtent: maxCrossExtent,
+        itemBuilder: itemBuilder,
+        childrenDelegate: SliverChildBuilderDelegate(
+          itemBuilder,
+          childCount: itemCount,
+          addAutomaticKeepAlives: addAutomaticKeepAlives ?? true,
+          addRepaintBoundaries: addRepaintBoundaries ?? true,
+          addSemanticIndexes: addSemanticIndexes ?? true,
+        ),
+      );
+
+  factory NikuGridView.custom({
+    Key? key,
+    Axis? scrollDirection,
+    bool? reverse,
+    ScrollController? controller,
+    bool? primary,
+    ScrollPhysics? physics,
+    bool? shrinkWrap,
+    EdgeInsetsGeometry? padding,
+    required SliverGridDelegate gridDelegate,
+    required IndexedWidgetBuilder itemBuilder,
+    int? itemCount,
+    bool? addAutomaticKeepAlives,
+    bool? addRepaintBoundaries,
+    bool? addSemanticIndexes,
+    double? cacheExtent,
+    int? semanticChildCount,
+    DragStartBehavior? dragStartBehavior,
+    ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior,
+    String? restorationId,
+    Clip? clipBehavior,
+  }) =>
+      NikuGridView(
+        type: NikuGridViewType.custom,
+        key: key,
+        scrollDirection: scrollDirection,
+        reverse: reverse,
+        controller: controller,
+        primary: primary,
+        physics: physics,
+        shrinkWrap: shrinkWrap,
+        padding: padding?.asNiku,
+        gridDelegate: gridDelegate,
+        itemBuilder: itemBuilder,
+        itemCount: itemCount,
+        addAutomaticKeepAlives: addAutomaticKeepAlives,
+        addRepaintBoundaries: addRepaintBoundaries,
+        addSemanticIndexes: addSemanticIndexes,
+        cacheExtent: cacheExtent,
+        semanticChildCount: semanticChildCount,
+        dragStartBehavior: dragStartBehavior,
+        keyboardDismissBehavior: keyboardDismissBehavior,
+        restorationId: restorationId,
+        clipBehavior: clipBehavior,
+      );
+
+  factory NikuGridView.extent({
+    Key? key,
+    Axis? scrollDirection,
+    bool? reverse,
+    ScrollController? controller,
+    bool? primary,
+    ScrollPhysics? physics,
+    bool? shrinkWrap,
+    EdgeInsetsGeometry? padding,
+    required double maxCrossAxisExtent,
+    double? mainAxisSpacing,
+    double? crossAxisSpacing,
+    double? childAspectRatio,
+    bool? addAutomaticKeepAlives,
+    bool? addRepaintBoundaries,
+    bool? addSemanticIndexes,
+    double? cacheExtent,
+    List<Widget> children = const [],
+    int? semanticChildCount,
+    DragStartBehavior? dragStartBehavior,
+    ScrollViewKeyboardDismissBehavior? keyboardDismissBehavior,
+    String? restorationId,
+    Clip? clipBehavior,
+  }) =>
+      NikuGridView(
+        type: NikuGridViewType.extent,
+        key: key,
+        scrollDirection: scrollDirection,
+        reverse: reverse,
+        controller: controller,
+        primary: primary,
+        physics: physics,
+        shrinkWrap: shrinkWrap,
+        padding: padding?.asNiku,
+        maxCrossAxisExtent: maxCrossAxisExtent,
+        mainAxisSpacing: mainAxisSpacing,
+        crossAxisSpacing: crossAxisSpacing,
+        childAspectRatio: childAspectRatio,
+        addAutomaticKeepAlives: addAutomaticKeepAlives,
+        addRepaintBoundaries: addRepaintBoundaries,
+        addSemanticIndexes: addSemanticIndexes,
+        cacheExtent: cacheExtent,
+        children: children,
+        semanticChildCount: semanticChildCount,
+        dragStartBehavior: dragStartBehavior,
+        keyboardDismissBehavior: keyboardDismissBehavior,
+        restorationId: restorationId,
+        clipBehavior: clipBehavior,
+      );
+
+  set count(int value) => itemCount = value;
+  set grid(int value) => itemCount = value;
+
+  set aspectRatio(double value) => childAspectRatio = value;
+  set aspect(double value) => childAspectRatio = value;
+  set ratio(double value) => childAspectRatio = value;
 
   set apply(NikuGridView? v) {
     if (v == null) return;
 
-    crossAxisCount = v.crossAxisCount != 1 ? v.crossAxisCount : crossAxisCount;
+    type = v.type ?? type;
+    itemCount = v.itemCount ?? itemCount;
+    children = v.children ?? children;
     scrollDirection = v.scrollDirection ?? scrollDirection;
     reverse = v.reverse ?? reverse;
     controller = v.controller ?? controller;
@@ -81,6 +316,12 @@ class NikuGridView extends StatelessWidget
         v.keyboardDismissBehavior ?? keyboardDismissBehavior;
     restorationId = v.restorationId ?? restorationId;
     clipBehavior = v.clipBehavior ?? clipBehavior;
+    itemBuilder = v.itemBuilder ?? itemBuilder;
+    maxCrossAxisExtent = v.maxCrossAxisExtent ?? maxCrossAxisExtent;
+    maxCrossExtent = v.maxCrossExtent ?? maxCrossExtent;
+    gridDelegate = v.gridDelegate ?? gridDelegate;
+    childrenDelegate = v.childrenDelegate ?? childrenDelegate;
+
     $internalParent..addAll(v.$internalParent);
   }
 
@@ -89,9 +330,10 @@ class NikuGridView extends StatelessWidget
   }
 
   NikuGridView get copied => NikuGridView(
-        crossAxisCount,
-        children,
         key: key,
+        type: type,
+        itemCount: itemCount,
+        children: children,
         scrollDirection: scrollDirection,
         reverse: reverse,
         controller: controller,
@@ -111,31 +353,114 @@ class NikuGridView extends StatelessWidget
         keyboardDismissBehavior: keyboardDismissBehavior,
         restorationId: restorationId,
         clipBehavior: clipBehavior,
+        itemBuilder: itemBuilder,
+        maxCrossAxisExtent: maxCrossAxisExtent,
+        maxCrossExtent: maxCrossExtent,
+        gridDelegate: gridDelegate,
+        childrenDelegate: childrenDelegate,
       )..$internalParent.addAll($internalParent);
 
-  GridView get widget => GridView.count(
-        key: key,
-        crossAxisCount: this.crossAxisCount,
-        children: this.children,
-        scrollDirection: this.scrollDirection ?? Axis.vertical,
-        reverse: this.reverse ?? false,
-        controller: this.controller,
-        primary: this.primary,
-        physics: this.physics,
-        shrinkWrap: this.shrinkWrap ?? false,
-        padding: this.padding?.value,
-        mainAxisSpacing: this.mainAxisSpacing ?? 0.0,
-        crossAxisSpacing: this.crossAxisSpacing ?? 0.0,
-        childAspectRatio: this.childAspectRatio ?? 1.0,
-        addAutomaticKeepAlives: this.addAutomaticKeepAlives ?? true,
-        addRepaintBoundaries: this.addRepaintBoundaries ?? true,
-        addSemanticIndexes: this.addSemanticIndexes ?? true,
-        cacheExtent: this.cacheExtent,
-        semanticChildCount: this.semanticChildCount,
-        dragStartBehavior: this.dragStartBehavior ?? DragStartBehavior.start,
-        keyboardDismissBehavior: this.keyboardDismissBehavior ??
-            ScrollViewKeyboardDismissBehavior.manual,
-        restorationId: this.restorationId,
-        clipBehavior: this.clipBehavior ?? Clip.hardEdge,
-      );
+  GridView get widget {
+    switch (type) {
+      case NikuGridViewType.builder:
+        return GridView.builder(
+          key: key,
+          scrollDirection: scrollDirection ?? Axis.vertical,
+          reverse: reverse ?? false,
+          controller: controller,
+          primary: primary,
+          physics: physics,
+          shrinkWrap: shrinkWrap ?? false,
+          padding: padding?.value,
+          gridDelegate: gridDelegate!,
+          itemBuilder: itemBuilder!,
+          itemCount: itemCount,
+          addAutomaticKeepAlives: addAutomaticKeepAlives ?? true,
+          addRepaintBoundaries: addRepaintBoundaries ?? true,
+          addSemanticIndexes: addSemanticIndexes ?? true,
+          cacheExtent: cacheExtent,
+          semanticChildCount: semanticChildCount,
+          dragStartBehavior: dragStartBehavior ?? DragStartBehavior.start,
+          keyboardDismissBehavior: keyboardDismissBehavior ??
+              ScrollViewKeyboardDismissBehavior.manual,
+          restorationId: restorationId,
+          clipBehavior: clipBehavior ?? Clip.hardEdge,
+        );
+
+      case NikuGridViewType.custom:
+        return GridView.custom(
+          key: key,
+          scrollDirection: scrollDirection ?? Axis.vertical,
+          reverse: reverse ?? false,
+          controller: controller,
+          primary: primary,
+          physics: physics,
+          shrinkWrap: shrinkWrap ?? false,
+          padding: padding?.value,
+          gridDelegate: gridDelegate!,
+          childrenDelegate: childrenDelegate!,
+          cacheExtent: cacheExtent,
+          semanticChildCount: semanticChildCount,
+          dragStartBehavior: dragStartBehavior ?? DragStartBehavior.start,
+          keyboardDismissBehavior: keyboardDismissBehavior ??
+              ScrollViewKeyboardDismissBehavior.manual,
+          restorationId: restorationId,
+          clipBehavior: clipBehavior ?? Clip.hardEdge,
+        );
+
+      case NikuGridViewType.extent:
+        return GridView.extent(
+          key: key,
+          scrollDirection: scrollDirection ?? Axis.vertical,
+          reverse: reverse ?? false,
+          controller: controller,
+          primary: primary,
+          physics: physics,
+          shrinkWrap: shrinkWrap ?? false,
+          padding: padding?.value,
+          maxCrossAxisExtent: maxCrossAxisExtent!,
+          mainAxisSpacing: mainAxisSpacing ?? 0.0,
+          crossAxisSpacing: crossAxisSpacing ?? 0.0,
+          childAspectRatio: childAspectRatio ?? 1.0,
+          addAutomaticKeepAlives: addAutomaticKeepAlives ?? true,
+          addRepaintBoundaries: addRepaintBoundaries ?? true,
+          addSemanticIndexes: addSemanticIndexes ?? true,
+          cacheExtent: cacheExtent,
+          children: children ?? [],
+          semanticChildCount: semanticChildCount,
+          dragStartBehavior: dragStartBehavior ?? DragStartBehavior.start,
+          keyboardDismissBehavior: keyboardDismissBehavior ??
+              ScrollViewKeyboardDismissBehavior.manual,
+          restorationId: restorationId,
+          clipBehavior: clipBehavior ?? Clip.hardEdge,
+        );
+
+      default:
+        return GridView.count(
+          key: key,
+          crossAxisCount: this.itemCount ?? 0,
+          children: this.children ?? [],
+          scrollDirection: this.scrollDirection ?? Axis.vertical,
+          reverse: this.reverse ?? false,
+          controller: this.controller,
+          primary: this.primary,
+          physics: this.physics,
+          shrinkWrap: this.shrinkWrap ?? false,
+          padding: this.padding?.value,
+          mainAxisSpacing: this.mainAxisSpacing ?? 0.0,
+          crossAxisSpacing: this.crossAxisSpacing ?? 0.0,
+          childAspectRatio: this.childAspectRatio ?? 1.0,
+          addAutomaticKeepAlives: this.addAutomaticKeepAlives ?? true,
+          addRepaintBoundaries: this.addRepaintBoundaries ?? true,
+          addSemanticIndexes: this.addSemanticIndexes ?? true,
+          cacheExtent: this.cacheExtent,
+          semanticChildCount: this.semanticChildCount,
+          dragStartBehavior: this.dragStartBehavior ?? DragStartBehavior.start,
+          keyboardDismissBehavior: this.keyboardDismissBehavior ??
+              ScrollViewKeyboardDismissBehavior.manual,
+          restorationId: this.restorationId,
+          clipBehavior: this.clipBehavior ?? Clip.hardEdge,
+        );
+    }
+  }
 }